+++ APITests/AutomationChecks/AutomationChecksTests.cs
@@ -35,7 +35,7 @@ public class AutomationChecksTests
             Name = "STT3: (the voices are back) vs (la planta)",
             NeedsAutoCheck = true,
             IsApiProcessed = true,
-            VerificationStatus = (int)MatchVerificationStatus.PendingVerification,
+            VerificationStatus = MatchVerificationStatus.PendingVerification,
             Tournament = tournament,
             TournamentId = tournament.Id
         };
